fstream: do not pass offset as part options member
offset was wrongly put into options.  It is something that will vary across
calls, like the file name, and not some property of the stream itself.

Some notes on the implementation: we have to get rid of the overload that
defines a buffer_size with a default value to avoid ambiguity with the version
that passes an offset and a default value for options. It would be great to
have offset and buffer size to be of guaranteed different types, but that won't
happen now.

Ideally, everybody should pass options. It is a more modern idiom and it seems
we are definitely moving in this direction. Inside seastar, it is possible to
get rid of the overload that passes buffer_size completely and always require
options to be passed. As a matter of fact, buffer size is only passed in tests,
and it seems to me that we don't really have a reason to pass non-standard
buffer sizes for those at all (but I'm keeping it as is at least for now).

But since we have out-of-tree users, I am providing a temporary overload that
allows for a position and a buffer size to be specified. After some time, it
should be possible to remove it entirely.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <1453168100-12077-1-git-send-email-glauber@scylladb.com>