rpc: add skip to simple_input_stream
This patch adds the ability to skip some bytes in the input stream
without reading them.

It is useful for situation when the reader reads an object with fields
it does not know how to read, but does know their size.

it also adds a method to return a simple_input_stream that is set to a
smaller size, which is usefull for an object that need to read only part
of the stream.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>
Message-Id: <1453308238-13414-1-git-send-email-amnon@scylladb.com>