fstream: improve make_file_input_stream() for a subrange of a file
Currently, when Scylla wants to read a subset of a file, it does so
by playing games with buffer_size and hoping no read-ahead takes place.

This is hacky, however, so provide an explicit API that specifies both
the offset within the file and the size of the data to read.  This will
allow us to enable read-ahead without fear.

Message-Id: <1453891684-3982-1-git-send-email-avi@scylladb.com>