Fix the do_marshall(sic) function to init_list
It is possible for the ignorer type to be optimized out with an
aggressive enough compiler (the calls likely not, but the object
construction is allowed, though no real impact).  Additionally, clang
warns on this object, since it is unused.

This patch removes the ignorer type (only used in 1 place) and replaces
the usage with a std::initializer_list that is cast to a void to prevent
causing an unused-value warning.

Signed-off-by: Erich Keane <erich.keane@verizon.net>
Message-Id: <1453939061-7695-4-git-send-email-erich.keane@verizon.net>