gate: add check() method
Usually, a gate allows stopping new invocations of some operation, and
waiting for those already started to complete. However, sometimes the
operation is very long, and when the gate is closed it would like to
know and finish prematurely.

So this patch adds a method gate::check() which checks if the gate is
closed, and if it is, it throws an exception (the same gate_closed_exception
that gate::enter() throws when the gate is closed). Long operations which
have natural places to stop (e.g., a loop of some sort) can place
gate::check() calls there, so they stop quickly when a gate is closed
(and allow gate::close() to resolve quickly).

This patch also adds an explanation, and an example, of gate::check()
to the tutorial.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1455530372-5111-1-git-send-email-nyh@scylladb.com>