read options from a secondary options file
In this patch, I am proposing options to be read from an additional file,
io.conf. The idea is that we keep the I/O options separate from the main file.

While I realize this can add some complexity, the main motivation to do that is
that this file will, in many cases, be autogenerated. If we go the append-only
route, running the configuration tool multiple times will leave many repeated
entries.  Running multiple times is not recommended, but there is no reason not
to protect against that. The best route is, then, to just write a new file in
its entirety. Writing the I/O options in a separate file just makes life a lot
easier.

Signed-off-by: Glauber Costa <glauber@scylladb.com>