apps: iotune: fix compilation with g++ 4.9
g++ 4.9 was giving the following error:

apps/iotune/iotune.cc:382:36: error: could not convert ‘{(uint64_t)IOPS, 1}’ from ‘<brace-enclosed initializer list>’ to ‘run_stats’
         return { uint64_t(IOPS), 1 };

This apparently happens because a compiler has interpreted the internal
fields initializes in 'run_stats' as an existence of explicit constructor(s)
which however didn't match the above aggregate initialization. This seems
to be a compiler bug and since this compilation error doesn't reproduce with
gcc 5.3.1 it must have been fixed.

In order to workaround this issue we'll define a 'run_stats' constructor
accepting 2 parameters that have default values (0).

Signed-off-by: Vlad Zolotarov <vladz@cloudius-systems.com>
Message-Id: <1455630157-19723-1-git-send-email-vladz@cloudius-systems.com>