reactor: new counters for the io queue
This patch introduces two new counters for each class of the I/O queue:
1) queue_length: we already have this metric globally, but in some cases it
   is useful to know the distribution of pending requests among classes
2) delay: if we are tracking down time related issues, we would like to know
   for how long is a request usually waiting to be serviced.

Note that "delay" is not a latency measure because it does not include the time
spent servicing the request itself. The main reason for that is so we don't
have to add an extra continuation to the I/O path, something that Avi opposed
(with good reasons) in the past. Moreover, we can have very good estimates
about that from OS disk tools (+- the seastar polling delay)

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <1456882128-11556-1-git-send-email-glauber@scylladb.com>