iotune: keep track of actual file_size, and bail on ENOSPC
We may not be able to write the whole file, and in cases like that, we would like to
dump a warning and move forward. So let's make file_size an instance member instead
of a static property of the iotune_manager. We can also use that to cut short our execution
in case of ENOSPC instead of using an early check.

Signed-off-by: Glauber Costa <glauber@scylladb.com>