shared_future: add companion class, shared_promise
It is not always the case we'll be getting a future from someone else and
wrapping it in a shared_future. It is a common idiom that we may be keeping
the promise ourselves. In that case, we would have to do some bookkeeping
that really is common to everybody using this idiom.

We can wrap that common code into a shared_promise.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <1458580502-5837-1-git-send-email-glauber@scylladb.com>