vla: Ensure memory is correctly freed
Structures with a variable length array returned from
make_struct_with_vla are deleted by the default destruction policy used
by std::unique_ptr. This means the sized global operator delete will be
called if the type of the structure is complete, to which the compiler
will pass the statically determined size of the structure; this will
most likely result in larger object being freed as smaller ones.

This patch fixes this by using malloc and free instead of the new and
delete operators.

Signed-off-by: Duarte Nunes <duarte@scylladb.com>
Message-Id: <1458691197-31331-2-git-send-email-duarte@scylladb.com>