tutorial: expand and improve semaphore section
Improve the organization of the semaphore section, and expand it from
what we learned in the last few days in ScyllaDB development of loops
with limited parallelism.

The section's exposition now starts with limiting of externally driven
events (where you cannot limit the events, but can start less work on
each) and only then progress to the more complex issues of limiting
of loops.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1458840689-15405-1-git-send-email-nyh@scylladb.com>