posix-stack: signal all errors as failed continuations
Currently, if there is a memory allocation failure during accept it
causes either accept() to throw an exception or return a failed future.
It makes it harder to handle such errors in accept() callers.
This patch changes the code so that all errors during accept are
signaled as a failed futures.

Signed-off-by: Paweł Dziepak <pdziepak@scylladb.com>