fair_queue: fix inconsistency during renormalization
When an accumulator is about to overflow, we renormalize.  Unforrunately
we do this inconsistently; a movement of +tau in time should divide the
accumulators by e, but here we divide the accumulators by FLT_MAX while
moving the time base by an arbitrary amount (moving it to steady_clock::now()).

The effect could be to squelch a queue completely for around tau time.

Fix by renormalizing consistently.  Since now renormalization doesn't
guarantee that cost will be finite, we renormalize in a loop.  With the
default parameters, the loop will run 10,000 times for each day of complete
idleness, or 3M times for each year of complete idleness.

Message-Id: <1460359952-30106-1-git-send-email-avi@scylladb.com>