smp: allow more than 128 in-flight operations on core-to-core queue
The 128-message limit on in-flight operations can lead to ABA deadlocks;
for example:

  A: send 128 requests to B
  B: while processing each of A's request, send a request back to A (processing
     A's request depends on completion of the B->A request).
  A: while processing the B->A requests, send a request back to B.

We deadlock. because A cannot send the last request to B.

See scylladb/scylla#1088 for a real-life example.

Fix by lifting the restriction; allow any number of requests to be
executed concurrently.  This requires the application to add its internal
limits, otherwise, if all shards simulataneously send many messages to a
single shard, that shard's memory can be exhausted.

Message-Id: <1461497600-20495-1-git-send-email-avi@scylladb.com>