reactor: kill max_inline_continuations
Since we have time quota based preemption, there is no need to count
continuations.  Continuation counting can undershoot, if continuations are
fast, resulting in excessive switching, or overshoot, if continuations are
slow, resulting in excessive latency (if we did not have the timer to
protect us).  Remove this accounting, simplifying the code and making it
more efficient.

Message-Id: <1462121442-17303-1-git-send-email-avi@scylladb.com>