thread: do not yield if idle
Currently, when we call yield into an idle system for a seastar::thread that
has an execution quota, we just leave the system idle until the period is over.
That's certainly not what we want most of the time, and although I could add a
flag to control that behavior, I doubt this is what we ever want. And if we do
want this existing behavior, that's the behavior that should be flagged, since
it really isn't the most obvious interpretation of "yield"

This patch changes to meaning of "yield" from the current "don't run me for
another x usec" to "give someone else a chance to run if I've run for more
than x usec", which as I have argued, it's the meaning it should've had all
along.

Signed-off-by: Glauber Costa <glauber@scylladb.com>

Message-Id: <1462315372-5168-1-git-send-email-glauber@scylladb.com>