iotune: don't coredump when directory fails to be created
The current file write code doesn't check if the directory was created
successfully because if it wasn't, writing the file itself will throw. However,
it is not always the case that creating a directory will return an error
message. It may throw, and if it does, the program will crash on an unhandled
exception. So let's change the code slightly so that everything related to file
creation falls within the try/catch block.

We do have to return an error code from this, which we currently don't do.

Fixes #129

Signed-off-by: Glauber Costa <glauber@scylladb.com>