scollectd: Support filtering specific collectd metrics
This patch adds the ability to turn on and off specific collectd
metrics.

It has the following changes:
add an enable flag to the collectd value_list it then uses that flag
when sending notification to decide if a specific meter should be sent
or not.

To be able to enable/disable the flag from the API two helper function
where added:

One that takes a bool and type_instance id and either enable or disable
that metrics.

and one that takes a type_instance_id and return if it's enable or
disable.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>
Message-Id: <1464525560-28242-2-git-send-email-amnon@scylladb.com>