rpc: wait for all replies to be completed before closing rpc server
After addition of resource accounting server is expected to be alive
until server handler is complete and reply is sent (this is when
resource is finally released), but current code destroys server to early
which may cause use after free during resource accounting. Fix that by
tracking outstanding handler with a gate and wait for a gate to be
closed in service's stop().

Message-Id: <1465290497-29562-3-git-send-email-gleb@scylladb.com>