util: add a seastar::value_of() helper function
seastar::value_of(Func) returns a value that may be used for evaluating
a value using the given functor.

The value is evaluated by a functor that gets no parameters.

The evaluation is triggered by operator() on a returned object.

In addition there is an implementation of a operator<< that would
allow printing a returned object as if it's a regular value.

For instance:

logger.debug("heavy eval result:{}", seastar::value_of([&] { return <heavy evaluation>; }));

(If a logging level is lower than "debug" the evaluation will not take place.)

Signed-off-by: Vlad Zolotarov <vladz@cloudius-systems.com>
Message-Id: <1466509402-6221-1-git-send-email-vladz@cloudius-systems.com>