resource: don't abort on too-high io queue count
When we added io queues, the number of io queues and cores were both
specified on the command line, so it made sense to abort on a conflict.

These days, the number of io queues is usually determined by iotune well
in advance, so the user has no idea what the error is or how to fix it.

Change the abort to a warning and silently adjust to avoid aborts.