reactor: run idle poll handler with a pure poll function
The idle work handler is called with a function to check whether the
idle handler needs to be preempted.  However, poll_once(), the function
actually used, is not pure: it can perform work on behalf of the user.
Examples are timers (invoked from the signal handler), or smp work_items.

This can cause surprises for the user code; either unexpected state changes
for the idle poll handler due to work performed in the poll function, or
unexpected state from the idle work handler impacting the tasks running in
poll_once().  The latter triggered scylladb/scylla#1395.

Our options to fix are to make poll_once() pure, or to introduce a
pure_poll_once() in parallel and use it from the idle handler only.
This patch implements the second method to avoid evaluating the performance
impact of moving work away from poll_once (specifically, the cache prefetches
used in smp).

Message-Id: <1467036869-28167-1-git-send-email-avi@scylladb.com>