app_template: Make run() wait for func when reactor exit is triggered externally
When reactor exit is triggered, for example on ^C, seastar will only
wait for exit hooks and pending continuations. It will not wait for
async function passed to run() to complete. It does wait for it until
interrupted though. In the normal non-interrupt path, when the
function exits by itself, it's likely releasing resources and does
some cleanup at the end of some execution chain. It's easier to
correctly implement cleanup on interrupt if the same path is used.

For example:

    bool interrupted = false;
    return app.run(argc, argv, [&] {
        // Could become engine().add_interrupt_handler() later, or
	// rely on default handler which would set engine().is_interrupted()
        engine().at_exit([&] () mutable { stop = true; return make_ready_future<>(); });

	auto deadline = now() + 3s;

        return do_until([&] { return interrupted || now() > deadline; }, [&] {
            return do_something();
        }).finally([&] {
	    return cleanup();
	});
    });

Message-Id: <1467728523-10033-1-git-send-email-tgrabiec@scylladb.com>