reactor: allow sleeping while disk aio is pending
Currently, the aio poller refuses to enter interrupt mode; this is usually
fine because the time to complete an I/O is often much less than the cpu
sleep/wake cycle.  But in overprovisioned environments, we might context
switch to another application if we have nothing to do but wait for aio.

This patch enables interrupt mode support by optionally having I/O
control blocks marked with an eventfd to wake when the I/O complete; we
also establish a dummy loop reading from this eventfd, so that the normal
epoll poller wakes up when it is ready.

Enabled by the new option --poll-aio 0, or by the existing option
--overprovisioned.

Message-Id: <1467973212-14723-1-git-send-email-avi@scylladb.com>