rpc: release send buffer earlier
Currently the code relies on buf.size() to know when non moved from
copy is destroyed which should trigger setting of a send completion
promise. This means though that buffer cannot be destroyed until send is
complete, but since sending of sstring copies this just waste memory.
Make promise<> to be optional instead and disengage it on move.