scollectd: Allow creating metric that is disabled
There are cases (specifically smp) that metrics are better be
unpublished by default.

To support that a varienet of the collectd helper function was added:
add_disabled_polled_metric.

Metrics that are created with this function will be disabled until they
would be set explicitly.

The add_polled helper function now gets an optional enabled param
(default to true) that state if the metrics is enable or disable.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>
Message-Id: <1468768626-19888-2-git-send-email-amnon@scylladb.com>