rpc: stop send loop even if it is stuck in send
Currently send loop is not aborted if it is stuck in _write_buf.write()
and receive loop exits on error. It will cause connection:stop() to
be stuck too. Fix this by shutting down writer side in stop_send_loop()
which will cause _write_buf.write() to exit. It is enough to shut down
only receiver in connection::stop() since this will cause receiver side
to exit and call stop_send_loop().