function_handler: Support general function handler
In most cases a function handler takes the request object and returns
either a json object or a future json object.

If the API need to return a future to a non json object more controll is
needed on the reply.

The general case, is that the handler gets a request and modify the
reply.

This patch adds support to register a general function handler.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>