file: emulate XFS locking
XFS will block in certain circumstances; the most common are that if an
operation changes the file size, it will block on any current AIO, and
may cause future AIO to block as well.  There are others, but this is the
most visible one.

This patch instates the same locking in seastar. Because seastar locks
are non-blocking, the reactor thread will not be blocked, while achieving
the same I/O performance.

As an intended side effect, the patch materializes blocked I/O operations
in a queue.  A future patch will be able to make use of the exposed queue
to further optimize things; this is the reason we use an explicit queue
and condition variable, rather than a reader/writer lock.

Operation ordering is maintained, to prevent starvation.