file: optimize xfs concurrent appending writes
XFS serializes and size-changing operation, not only against other
size-changing operations, but also against other reads and writes.

append_challenged_posix_file_impl recognizes this and also serializes
size-changing operations to avoid XFS blocking the reactor thread.  At
this point, we can look ahead into our operation queue and truncate the
file so that all the size-changing operations we are about to issue will
become non-size-changing operations; XFS can then process them in parallel.

This requires that ftruncate() not block, but it indeed does not, if no
other operations are in progress.