file: allow files with sloppy size, in order to work around XFS locking constraints
XFS likes to serialize I/O when a file's size is changed.  This slows down
append workloads.

To work around this, introduce a sloppy_size file option.  When the option
is enabled, we extend the file size speculatively to a large size (doubling
each time).  Until we hit the speculative size limit, append operations are
actually non-size-extending, and can proceed in parallel.

Since the on-disk file size is now speculative, we introduce a variable
to track the logical file size, and truncate the file back when it is
flushed or closed.