simple_input_stream: introduce copy_to() member function
Currently, if a user of simple_input_stream want to efficiently
the input stream to some output stream it has to use begin() and
size(). This interface is not general enough, since it assumes
that the underlying buffer is continuous.

This patch introduces copy_to() function that takes output
stream as an argument and copies to it the data from this stream.

Signed-off-by: Paweł Dziepak <pdziepak@scylladb.com>
Message-Id: <1470786434-23721-1-git-send-email-pdziepak@scylladb.com>