reactor: fix I/O queue pending requests collectd metric
Tomek has recently seen a SIGSEGV happening in our collectd subsystem. He
describes the problem, a consequence of our call to queued_requests() as
follows:

"_io_queue may point to an object which is owned and mutated by a different
shard (_io_coordinator). io_queue::queued_requests() calls
fair_queue::waiters(), which calls semaphore::waiters, which calls
chunked_fifo::size(), which is not thread-safe. Which suggests that
queued_requests() should only be called on _io_coordinator, but it's not."

This patch changes the collectd registration so that the counter is only
registered in the correct shards.

Fixes #177

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <1470830825-27885-1-git-send-email-glauber@scylladb.com>