threads: yield on get if we had run for too long.
Currently, we keep chaining continuations until need_preempt() returns true.
But when it comes to thread, we just run for as long as we can, until the thread
user explicitly calls yield.

That is suboptimal, because unless the caller code sprinkles yields all over the
place, we can never guarantee that we are yielding often enough.

My proposal to fix this is to embed the check in natural actions that will be
taken by the thread callers anyway. The one thing that distinguishes seastar
threads from the rest of our future chains is that we synchronously call get()
in futures. We know for a fact that at this point the thread may wait already -
if it turns out that the future is not ready. So why not always yield at this
point (if we have been running for this long), even if the future is available?

Why are we yielding instead of just calling wait() if need_preempt() is a valid
question. The reason is that threads that belong to a scheduling group are
expected to have different behavior in this case, and should only run for so
long. If we were to just wait() at this point, those threads would just
switch_out and then switch_in back at the next opportunity - which defeats the
purpose of their scheduling group.

Generally speaking, for the existing attributes and conceivably for the ones to
come, the desired behavior of the thread is encoded in its yield method - so
that's what we should call.

Of course, if the thread does not call get() often enough then it might have to
yield explicitly. But for the majority of the thread uses, this offers a
transparent and efficient way of controlling runtime.

Before this patch, I can see often the  ScyllaDB memtable writer - which runs on
a thread - running continuously for 8 - 10ms. After this patch, we never run for
more than task_quota_ms.

Signed-off-by: Glauber Costa <glauber@scylladb.com>