move need_preemt to its own file
This is to avoid circular dependencies between future.hh and thread_impl.hh.
One way to solve this would be to simply move need_preempt() to thread_impl.hh -
and let future.hh just import from it.

But I believe it is cleaner to have this into a separate file that can be
imported from both.

Signed-off-by: Glauber Costa <glauber@scylladb.com>