threads: calculate yield point when we switch in to a new thread
We currently have a should_yield function that we could very well expose through
the thread_impl mechanism. However, it would be great to be able to decide
whether or not to yield without calling into external functions. This patch adds
a field to the jmp_buf_link that tracks when are we supposed to yield (if we do
time based yielding). A caller deciding whether or not to yield can just check
the current timestamp against the expected timestamp and yield.

We may still have external calls for the timestamp, but that is less of a
concern.  First, because get() is now already inline, this will only be called
for threads.  Secondly, because we check whether or not the time field (yield
at) exists in the first place, we do this only for threads in a scheduling
group.

But more importantly, if we have a local clock that uses the timestamp counter
(in my todo list), then we won't be paying an external call price not even in
this case.

Signed-off-by: Glauber Costa <glauber@scylladb.com>