file: ext4 support
It turns out that ext4 works if we don't change sizes, so the write-behind
work can be adapted to support ext4 by calling ftruncate on any size-
changing write (not just if there is more than one of them).  There is
some magic involved in determining the filesystem type, but not too much.