file: improve tmpfs support
Currently we support tmpfs if --relaxed-dma is provided, but we can do
better: while tmpfs does not support O_DIRECT, neither will it block
during io_submit() (it will copy the data immediately, but that is
non-blocking behavior; at least if swap is not involved).

We can improve tmpfs support by detecting it, and forgiving the lack
of O_DIRECT if we're indeed on tmpfs.  This helps in testing environments.


Reviewed-by: Raphael S. Carvalho <raphaelsc@scylladb.com>
Message-Id: <1472290990-17697-1-git-send-email-avi@scylladb.com>