metrics: Add the metrics API and implementation
There are two kind of users of the metrics layer. Mainly, classes that
register metrics, that uses the metrics_registration.hh and metrics.hh
that holds the logic for registering a metrics.

Second, there are APIs (collectd, prometheus, REST Api) that need
access to the metrics internal, they can use the metrics_api.hh

The metrics layer implementation is based on a map between an id
and a pointer to registered_metric.

registered_metric is a base class, impl_registered_metric that inherits
from it, is a template class that holds a reference to a function, that
return metric_value.

The metric_value is a class that holds the different values supported by
the metrics layer.

Registering a metric is done by calling add_group with a list of
metrics.

so if the _metrics is metric_groups
This code will add two gauge:

_metrics->add_group("cache", {
  sm::make_gauge("bytes", "used", [this] { return
    _region.occupancy().used_space(); }),
  sm::make_gauge("bytes", "free", [this] { return
    _region.occupancy().free_space(); })
});

The implementation accept either reference or function as value by
wrapping reference to a value in a lambda function.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>