semaphore: require explict setting for start value
Our semaphore currently defaults, if no value is explicitly given, to
start with a value of 1. This is useful for some scenarios, but wrong
for others, and it's all to easy to forget to explicitly specify it
or assume it defaults to something else (like 0). We recently had two
bugs (the latest one in condition_variable) because of exactly this
assumption.

Even in the rare cases where a start value of 1 is desired, it's better
to explicitly mention it instead of relying on this be the default.

So let's just drop this default.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1474460254-20300-1-git-send-email-nyh@scylladb.com>