iostream: Fix stack overflow in output_stream::split_and_put()
When the buffer passed to output_stream::write() is large and the
output stream is setup to trim buffers to certain size before handing
over to the sink, split_and_put() may have to issue a lot of puts()
before it completes. We may recurse a lot without deferring if the
undrelying data sink is fast or allows a lot of data to be
buffered. That in turn may lead to stack overflow.

Fix by avoiding recursion.

That seems to be the case with ScyllaDB's output_stream of the sstable
data file writer when writing large cell values. The outer-most stream
splits buffers into 4K chunks, which are then compressed by
compressed_file_data_sink_impl, which then hands them over to the
file's output stream. The latter stream allows 10 x 128K of data to be
buffered (write-behind), issuing IO in the background. It means that
320 (10x128/4) or more split_and_put() iterations could proceed
without deferring. Each iteration takes 448 bytes on stack. That gives
us 143K, which is more than seastar thread's stack size.

This manifested itself as a heap memory corruption (seastar thread's
don't have stack guards) leading to a peculiar crash.

Refs scylladb/scylla#1622

Message-Id: <1474470637-20758-1-git-send-email-tgrabiec@scylladb.com>