core: Print backtrace on SIGSEGV and SIGABRT
Often when seastar application terminates on assertion failure or
segmentation fault there is not enough information to figure out where
the problem is. Knowing the backtrace can be enough to identify the
problem. This change causes seastar to dump backtrace on such events
to make addressing user reports easier.

Raw addresses are currently dumped. It would be possible to print
symbol names, but currently we ship stripped programs and in that case
we would only get misleading symbol names.

Example output:

ERROR 2016-09-23 16:56:47,816 [shard 0] seastar - Failed to allocate 5534 bytes
Aborting on shard 0.
Backtrace:
  0x00000000004893ba
  0x000000000048949b
  0x00007f0f806f4a00
  0x00007f0f80357a28
  0x00007f0f8035962a
  0x000000000051acd4
  0x000000000051ad50
  0x000000000051aed3
  0x000000000079b295
  0x0000000000839668
  0x00000000008396a9
  0x0000000000c53490
  0x0000000000c8bafc
  0x0000000000c55915
  0x0000000000c575f5
  0x0000000000acb656
  0x0000000000acbc88
  0x000000000048c22f
  0x00000000004c6b83
  0x0000000000543df4
  0x000000000041deab
  0x00007f0f80343580
  0x0000000000487819