prometheus: prevent illegal prometheus names
Not all the collectd metrics, has valid names, some of them contains
spaces and minus sign.

When used with the collectd_exporter, the collectd_exporter clean those
names out, but when we export the name directly, the prometheus server
complains about it.

Note that this is temporary, the clean up that is schedule to be done on
the collectd metrics will remove the need for this function.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>
Message-Id: <1475132488-2740-3-git-send-email-amnon@scylladb.com>