fstream: avoid issuing reads in batches
Original read ahead implementation issued new reads in two cases:
 - get() was called an there was no ready or in progress read
 - a read has completed and the number of ready of in progress reads
   was below configured read ahead threshold

That's not really how we would like read ahead to work. If all buffers
queued in _read_buffers are ready no new reads will be issued until
all of them are consumed. This is not a desired behaviour since it
doesn't hide latency as well as it could.

The solution is to change the code so that more reads are issued
only when a buffer is consumed.

Signed-off-by: Paweł Dziepak <pdziepak@scylladb.com>