build: Work around for libunwind messing up signal mask
The libunwind library version which ships with Ubuntu 16.04
(libunwind-8.0.1) has a bug which causes signal mask to be undefined
after exceptions are thrown.

See https://savannah.nongnu.org/bugs/?48486.

This manifests in a number of ways. One is that some operations appear
to hang due to worker thread completion notifications not being
delivered. Collectd statistics are not exported due to timers not
firing. It may also cause aborts on SIGTERM if it's delivered to
non-0th shard.

Refs scylladb/scylla#1725

The workaround puts libgcc_s before libunwing on the link list, so
that gcc's default unwinding handlers are used instead of those from
libunwind.