collectd: expose the hostname
The collectd uses either a configurable hostname or if no configuration
is found the local host name when exposing its matrics.

The prometheus should use the same value to mark its instance so metrics
comming from collectd and metrics comming from prometheus would look the
same.

This patch adds a method that export the host used by collectd so it can
be used by prometheus.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>