fstream: add dynamic adjustments based on stream history
This patch allows the user to provide fstream with a pointer to the
stream histroy which would enable dynamic adjustments of buffer size.

The stream uses the history to determine whether or not enter slow start
phase in which it uses smaller buffer sizes in order to minimise wasted
IO bandwidth in case the read data is dropped.

Signed-off-by: Paweł Dziepak <pdziepak@scylladb.com>