scollectd: Change the scollectd to use the metrics API
The metircs layer gives a cleaner API for registering metrics,
eventually metrics will be register using the metrics API, but
currently, the collectd API is used throughout the system.

This patch is for the transition phase. It changes the collectd to use
the metrics API to get it's value. For the current code to continue
working it uses the collectd registration method to register the metrics
on the metrics layer.

When all the registration in the system will be replaced to use the
metrics layer, the collectd will be cleaned removing all the
registration related code from it.