rpc: Add missing adjustment of snd_buf::size
The code was trimming the buffer, but snd_buf::size was not
adjusted. As a result, the value of uncompressed message size which
was written to the frame will not match the uncompressed size which
was passed to LZ4_compress. lz4_compressor::compress() passes the size
of the temporary_buffer to LZ4_compress but writes snd_buf::size into
the frame. This will cause failure in LZ4_decompress_fast() on the
other side:

  rpc - client 10.154.2.97: server connection dropped: RPC frame LZ4 decompression failure

We hit this in a mixed ScyllaDB cluster with some nodes being on
1.3.1 and some on 1.4-rc2. See scylladb/scylla#1767.

Message-Id: <1476892857-4907-1-git-send-email-tgrabiec@scylladb.com>