future: prioritise continuations that can run immediately
Currently, code like this:

    return do_some_io().then([] {
        // ...
        return something_that_returns_ready_future().then([] {
            // ...
            return something_that_returns_ready_future();
        }).then([] {
            // ...
            return something_that_returns_ready_future();
        });
    });

is more efficient than this:

    return do_some_io().then([] {
        // ...
        return something_that_returns_ready_future();
    }).then([] {
        // ...
        return something_that_returns_ready_future();
    }).then([] {
        // ...
        return something_that_returns_ready_future();
    });

The reason for that is that the in the former case the continuations
that wait for the ready future can be immediately executed without
allocating task object and scheduling them in the reactor queue.
That's not possible in the latter case as the first continuation has
to wait for the IO to complete and, consequently, the subsequent
continuations cannot benefit from the fact that they depend on the
futures that are already ready. This introduces more deferring points
which increases latency and ruins cache locality.

The solution introduced in this patch is to make sure that if the
future returned by the continuation is already ready and there is
another continuation chained to it that continuation is executed
immediately. This means that while the latter example still involves
more allocations, there are no additional deferring and the cache
locality is better.

Signed-off-by: Paweł Dziepak <pdziepak@scylladb.com>
Message-Id: <1477381025-32083-1-git-send-email-pdziepak@scylladb.com>