rpc: Silence warning about ignored failed future
Can be observed, for instance, when replicas start to timeout readers
due to disk not keeping up.

Example log:

  WARN  2016-10-24 20:02:40,225 [shard 0] seastar - Exceptional future ignored: semaphore_timed_out (Semaphore timedout)

The future is ignored because in this path response is not being sent,
because the client is marked as being in error, which means we're
closing the connection or have already done so.

Fixes #176

Message-Id: <1477382392-6370-1-git-send-email-tgrabiec@scylladb.com>