reactor: improve task quota timer resolution
Commit 4942016fb914ea46a974f7b3747e694f3df4cfe7 ("reactor: switch task quota
timer to CLOCK_THREAD_CPUTIME_ID") switched the task quota timer to a
cputime clock, to avoid counting the time we spend sleeping, as this results
in spurious wakes during sleep, and therefore high idle cpu consumption.
However, an uninteneded side effect occured since CLOCK_THREAD_CPUTIME_ID has
1/HZ resolution in Linux, rather than being a high resolution timer.

With the default configuration, on a Red Hat kernel that doesn't make much
of a difference, since HZ=1000 and the task quota is 2 ms.  But on Ubuntu
kernels (where HZ=250 -> 4ms resolution), or if someone sets --task-quota-ms,
we have the task quota timer having lower resolution than the task quota,
so we consistently miss it.  This hurts latency sensitive loads.

To fix this, change back to the high resolution CLOCK_MONOTONIC clock, but
disable it while we sleep to avoid spurious wakeups.

Message-Id: <1477572725-10444-1-git-send-email-avi@scylladb.com>